pcd: add stage_values_matches_wire_count tests for all nested stages
Move assert_stage_values helper from native/stages to shared
internal/tests so both sides can use it. Embed the test in
define_bridge_stage! so every macro invocation gets coverage
automatically, and add the same test to the hand-written
nested preamble.
