+++ crates/ragu_pcd/src/internal/nested/stages/preamble.rs
@@ -143,3 +143,15 @@ impl<C: CurveAffine, R: Rank> ragu_circuits::staging::Stage<C::Base, R> for Stag
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::internal::tests::{R, assert_stage_values};
+    use ragu_pasta::EqAffine;
+
+    #[test]
+    fn stage_values_matches_wire_count() {
+        assert_stage_values(&Stage::<EqAffine, R>::default());
+    }
+}
